% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-bootstrapped.iRAM.2node.R
\docType{data}
\name{bootstrap.iRAM.2node}
\alias{bootstrap.iRAM.2node}
\title{Bootstrapped iRAM (including replications of iRAM and corresponding time profiles)}
\format{An object of class \code{list} of length 5.}
\usage{
bootstrap.iRAM.2node
}
\description{
Bootstrapped iRAM (including replications of iRAM and corresponding time profiles)
}
\details{
Data bootstrapped from the estimated three-node network structure with 200 replications.
}
\examples{
data(bootstrap.iRAM.2node)
\dontshow{
bootstrap.iRAM.2node$mean # mean of bootstrapped iRAM
bootstrap.iRAM.2node$upper # Upper bound of confidence interval of bootstrapped iRAM
bootstrap.iRAM.2node$lower # lower bound of confidence interval of bootstrapped iRAM
bootstrap.iRAM.2node$time.profile.data # time profiles generated from the bootstrapped beta matrices
bootstrap.iRAM.2node$recovery.time.reps # iRAMs generated from the bootstrapped beta matrices
}
\donttest{
bootstrap.iRAM.2node$mean # mean of bootstrapped iRAM
bootstrap.iRAM.2node$upper # Upper bound of confidence interval of bootstrapped iRAM
bootstrap.iRAM.2node$lower # lower bound of confidence interval of bootstrapped iRAM
bootstrap.iRAM.2node$time.profile.data # time profiles generated from the bootstrapped beta matrices
bootstrap.iRAM.2node$recovery.time.reps # iRAMs generated from the bootstrapped beta matrices
}
}
\keyword{datasets}
