% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-bootstrapped_iRAM_2node.R
\docType{data}
\name{bootstrap_iRAM_2node}
\alias{bootstrap_iRAM_2node}
\title{Bootstrapped iRAM (including replications of iRAM and corresponding time profiles) for the bivariate time-series (simts2node)}
\format{An object of class \code{list} of length 5.}
\usage{
bootstrap_iRAM_2node
}
\description{
Bootstrapped iRAM (including replications of iRAM and corresponding time profiles) for the bivariate time-series (simts2node)
}
\details{
Data bootstrapped from the estimated three-node network structure with 200 replications.
}
\examples{
\dontshow{
bootstrap_iRAM_2node$mean # mean of bootstrapped iRAM
bootstrap_iRAM_2node$upper # Upper bound of confidence interval of bootstrapped iRAM
bootstrap_iRAM_2node$lower # lower bound of confidence interval of bootstrapped iRAM
bootstrap_iRAM_2node$time.profile.data # time profiles generated from the bootstrapped beta matrices
bootstrap_iRAM_2node$recovery.time.reps # iRAMs generated from the bootstrapped beta matrices
}
\donttest{
bootstrap_iRAM_2node$mean # mean of bootstrapped iRAM
bootstrap_iRAM_2node$upper # Upper bound of confidence interval of bootstrapped iRAM
bootstrap_iRAM_2node$lower # lower bound of confidence interval of bootstrapped iRAM
bootstrap_iRAM_2node$time.profile.data # time profiles generated from the bootstrapped beta matrices
bootstrap_iRAM_2node$recovery.time.reps # iRAMs generated from the bootstrapped beta matrices
}
}
\keyword{datasets}
