% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse_response.R
\name{compute.recovery.time.from.beta}
\alias{compute.recovery.time.from.beta}
\title{Generate recovery time (impulse response anlaysis metric) from beta matrix.}
\usage{
compute.recovery.time.from.beta(beta.matrix, var.number, lag.order = 1,
  threshold = 0.01, replication = 200, steps = 100)
}
\arguments{
\item{beta.matrix}{beta matrix}

\item{var.number}{number of variables in the time series}

\item{lag.order}{lag order of the model to be fit}

\item{threshold}{threshold of calculation of recovery time (duration of perturbation), default value is 0.01}

\item{replication}{number of replication of bootstrap, default value is 200}

\item{steps}{number of steps of impulse response analysis, default value is 100}
}
\value{
iRAM matrix. Rows represent where the orthognal impulse was given, and columns represent the response. Dimension is var.number by var.number.
}
\description{
Generate recovery time (impulse response anlaysis metric) from beta matrix.
}
