% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impulse_response.R
\name{iRAM}
\alias{iRAM}
\title{Generate iRAM (impulse response anlaysis metric) from model fit.}
\usage{
iRAM(model.fit, var.number, lag.order = 1, threshold = 0.01, boot = FALSE,
  replication = 200, steps = 100)
}
\arguments{
\item{model.fit}{model fit object generated by lavaan}

\item{var.number}{number of variables in the time series}

\item{lag.order}{lag order of the model to be fit}

\item{threshold}{threshold of calculation of recovery time (duration of perturbation), default value is 0.01}

\item{boot}{to bootstrap, default value is FALSE}

\item{replication}{number of replication of bootstrap, default value is 200}

\item{steps}{number of steps of impulse response analysis, default value is 100}

\item{plot.histogram}{to plot distribution of iRAM, default value is FALSE}
}
\value{
iRAM matrix. Rows represent where the orthognal impulse was given, and columns represent the response. Dimension is var.number by var.number.
}
\description{
Generate iRAM (impulse response anlaysis metric) from model fit.
}
