% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-model.fit.R
\docType{data}
\name{model.fit}
\alias{model.fit}
\title{Model fitbased on similated time-series by uSEM.}
\format{An object of class \code{lavaan} of length 1.}
\usage{
model.fit
}
\description{
Model fitbased on similated time-series by uSEM.
}
\examples{

\dontshow{
data(model.fit)
 }
\donttest{
data(model.fit)
 }

}
\keyword{datasets}
