% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSEM.R
\name{model.summary}
\alias{model.summary}
\title{Provide model summary. Model fit criteria: 3 out of 4 rule, meaning 3 out of 4 critea should be satisfied, including CFI and NFI should be greater than 0.9, RMSEA and SRMR should be less than 0.08.}
\usage{
model.summary(model.fit, var.number, lag.order)
}
\arguments{
\item{model.fit}{model fit object generated by lavaan}

\item{var.number}{number of variables in the time-series}

\item{lag.order}{lag order of model}
}
\value{
beta matrix estimates

matrix of standard error of beta

matrix of psi estimates

fit statistics CFI

fit statistics TLI

fit statistics RMSEA

fit statistics SRMR
}
\description{
Provide model summary. Model fit criteria: 3 out of 4 rule, meaning 3 out of 4 critea should be satisfied, including CFI and NFI should be greater than 0.9, RMSEA and SRMR should be less than 0.08.
}
