% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSEM_20171004.R
\name{parse.beta}
\alias{parse.beta}
\title{Parse the beta from model fit object}
\usage{
parse.beta(var.number, model.fit, lag.order, matrix = F)
}
\arguments{
\item{var.number}{number of variables in the time series}

\item{model.fit}{model fit object generated by lavaan}

\item{lag.order}{lag order of the model to be fit}

\item{matrix}{output beta in matrix format or estimates format, default value is FALSE (as estimates)}
}
\value{
beta
}
\description{
Parse the beta from model fit object
}
