% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.iRAM.dist}
\alias{plot.iRAM.dist}
\title{Plot distribution of recovery time based on bootstrapped version of iRAM}
\usage{
\method{plot}{iRAM.dist}(recovery.time.reps)
}
\arguments{
\item{recovery.time.reps}{bootstrapped version of recovery time}
}
\value{
a ggplot object
}
\description{
Plot distribution of recovery time based on bootstrapped version of iRAM
}
