% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.network.graph}
\alias{plot.network.graph}
\title{Plot the network graph}
\usage{
\method{plot}{network.graph}(beta, var.number)
}
\arguments{
\item{beta}{matrix of temporal relations, cotaining both lag-1 and contemporaneous}

\item{var.number}{number of variables in the time series}
}
\description{
Plot the network graph
}
