% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.time.profile}
\alias{plot.time.profile}
\title{Plot time profiles given a time-series generated by impulse response analysis}
\usage{
\method{plot}{time.profile}(time.series.data, threshold, xupper = 20)
}
\arguments{
\item{time.series.data}{data of impulse response in long format}

\item{threshold}{threshold of asymptote of equilibrium}

\item{xupper}{upper limit of x-axis}
}
\description{
Plot time profiles given a time-series generated by impulse response analysis
}
