% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_integrated_time_profile}
\alias{plot_integrated_time_profile}
\title{Plot the time profiles in the integrated form}
\usage{
plot_integrated_time_profile(beta.matrix, var.number, lag.order = 1)
}
\arguments{
\item{beta.matrix}{matrix of temporal relations, cotaining both lag-1 and contemporaneous}

\item{var.number}{number of variables in the time series}

\item{lag.order}{lag order of the model to be fit}
}
\description{
Plot the time profiles in the integrated form
}
\examples{
\dontshow{
plot_integrated_time_profile(beta.matrix = true_beta_3node,
                  var.number = 3,
                  lag.order = 1)
}

\donttest{
plot_integrated_time_profile(beta.matrix = true_beta_3node,
                  var.number = 3,
                  lag.order = 1)
}

}
