% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_network_graph}
\alias{plot_network_graph}
\title{Plot the network graph}
\usage{
plot_network_graph(beta, var.number)
}
\arguments{
\item{beta}{matrix of temporal relations, cotaining both lag-1 and contemporaneous}

\item{var.number}{number of variables in the time series}
}
\description{
Plot the network graph
}
