% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_time_profile}
\alias{plot_time_profile}
\title{Plot time profiles given a time-series generated by impulse response analysis}
\usage{
plot_time_profile(time.series.data, var.number, threshold = 0.01,
  xupper = 20)
}
\arguments{
\item{time.series.data}{data of impulse response in long format}

\item{var.number}{number of variables in the time-series}

\item{threshold}{threshold of asymptote of equilibrium}

\item{xupper}{upper limit of x-axis}
}
\description{
Plot time profiles given a time-series generated by impulse response analysis
}
\examples{
\dontshow{
data(bootstrap.iRAM.2node)

plot_time_profile(time.series.data = bootstrap.iRAM.2node$time.profile.data,
                 var.number = 2,
                 threshold= .01,
                 xupper = 20)
}

\donttest{
data(bootstrap.iRAM.2node)

plot_time_profile(time.series.data = bootstrap.iRAM.2node$time.profile.data,
                 var.number = 2,
                 threshold= .01,
                 xupper = 20)
}

}
