% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-simts2node.R
\docType{data}
\name{simts2node}
\alias{simts2node}
\title{Simulated time-series data}
\format{An object of class \code{data.frame} with 200 rows and 2 columns.}
\usage{
simts2node
}
\description{
Simulated time-series data
}
\details{
Data simulated from a given three-node network structure with 200 measurements.
        Network structure is shown in the dataset true.beta.
        Process noise has mean of 0 and SD .1.
}
\keyword{datasets}
