% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-simts_2node.R
\docType{data}
\name{simts_2node}
\alias{simts_2node}
\title{Simulated bivariate time-series data}
\format{An object of class \code{data.frame} with 200 rows and 2 columns.}
\usage{
simts_2node
}
\description{
Simulated bivariate time-series data
}
\details{
Data simulated from a given three-node network structure with 200 measurements.
        Network structure is shown in the dataset true.beta.
        Process noise has mean of 0 and SD .1.
}
\examples{
\dontshow{
data(simts_2node)
}
\donttest{
data(simts_2node)
}

}
\keyword{datasets}
