% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-simts_3node.R
\docType{data}
\name{simts_3node}
\alias{simts_3node}
\title{Simulated 3-variate time-series data}
\format{An object of class \code{data.frame} with 100 rows and 3 columns.}
\usage{
simts_3node
}
\description{
Simulated 3-variate time-series data
}
\details{
Data simulated from a given three-node network structure with 200 measurements.
        Network structure is shown in the dataset true.beta.
        Process noise has mean of 0 and SD .1.
}
\examples{
\dontshow{
data(simts_3node)
 }
\donttest{
data(simts_3node)
 }

}
\keyword{datasets}
