% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-true.beta.2node.R
\docType{data}
\name{true.beta.2node}
\alias{true.beta.2node}
\title{The true beta matrix (4 by 4) used in simulation.}
\format{An object of class \code{matrix} with 4 rows and 4 columns.}
\usage{
true.beta.2node
}
\description{
The true beta matrix (4 by 4) used in simulation.
}
\details{
true.beta.2node <-  matrix(c(0,0,0,0,
                                0,0,0,0,
                               0.2,-.4,0,-0.25,
                               0,0.3,-0.2,0),
                     nrow = 4,
                     ncol = 4,
                     byrow = TRUE)
}
\examples{
data(true.beta.2node)
\dontshow{
true.beta.2node
 }
\donttest{
true.beta.2node
 }

}
\keyword{datasets}
