% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-true.beta.R
\docType{data}
\name{true.beta}
\alias{true.beta}
\title{The true beta matrix (6 by 6) used in simulation
true.beta <-  matrix(c(0,0,0,0,0,0,
                      0,0,0,0,0,0,
                      0,0,0,0,0,0,
                      0.2,0,0.25,0,0,0.6,
                      0,0.3,0,-0.2,0,-0.6,
                      0,-0.2,0.3,0,0,0),
                     nrow = 6,
                     ncol = 6,
                     byrow = TRUE)}
\format{An object of class \code{matrix} with 6 rows and 6 columns.}
\usage{
true.beta
}
\description{
The true beta matrix (6 by 6) used in simulation
true.beta <-  matrix(c(0,0,0,0,0,0,
                      0,0,0,0,0,0,
                      0,0,0,0,0,0,
                      0.2,0,0.25,0,0,0.6,
                      0,0.3,0,-0.2,0,-0.6,
                      0,-0.2,0.3,0,0,0),
                     nrow = 6,
                     ncol = 6,
                     byrow = TRUE)
}
\examples{
data(true.beta)
\donttest{
true.beta
 }
}
\keyword{true.beta}
