% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-true_beta_2node.R
\docType{data}
\name{true_beta_2node}
\alias{true_beta_2node}
\title{The true beta matrix (4 by 4) used in simulation.}
\format{An object of class \code{matrix} with 4 rows and 4 columns.}
\usage{
true_beta_2node
}
\description{
The true beta matrix (4 by 4) used in simulation.
}
\details{
true_beta_2node <-  matrix(c(0,0,0,0,
                                0,0,0,0,
                               0.2,-.4,0,-0.25,
                               0,0.3,-0.2,0),
                     nrow = 4,
                     ncol = 4,
                     byrow = TRUE)
}
\examples{
\dontshow{
true_beta_2node
 }
\donttest{
true_beta_2node
 }

}
\keyword{datasets}
