% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-true_beta_3node.R
\docType{data}
\name{true_beta_3node}
\alias{true_beta_3node}
\title{The true beta matrix (6 by 6) used in simulation.}
\format{An object of class \code{matrix} with 6 rows and 6 columns.}
\usage{
true_beta_3node
}
\description{
The true beta matrix (6 by 6) used in simulation.
}
\details{
true_beta_3node <-  matrix(c(0,0,0,0,0,0,
                      0,0,0,0,0,0,
                      0,0,0,0,0,0,
                      0.2,0,0.25,0,0,0.6,
                      0,0.3,0,-0.2,0,-0.6,
                      0,-0.2,0.3,0,0,0),
                     nrow = 6,
                     ncol = 6,
                     byrow = TRUE)
}
\examples{
\dontshow{
true_beta_3node
 }
\donttest{
true_beta_3node
 }

}
\keyword{datasets}
