% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-usemmodelfit.R
\docType{data}
\name{usemmodelfit}
\alias{usemmodelfit}
\title{Model fitbased on similated time-series by uSEM.}
\format{An object of class \code{lavaan} of length 1.}
\usage{
usemmodelfit
}
\description{
Model fitbased on similated time-series by uSEM.
}
\examples{

\dontshow{
data(usemmodelfit)
 }
\donttest{
data(usemmodelfit)
 }

}
\keyword{datasets}
