% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uSEM.R
\name{W2E}
\alias{W2E}
\title{Convert the matrix format of edges into the list format of edges}
\usage{
W2E(x)
}
\arguments{
\item{x}{matrix of temporal relations, cotaining both lag-1 and contemporaneous}
}
\value{
edge list
}
\description{
Convert the matrix format of edges into the list format of edges
}
