% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_iRAM_dist}
\alias{plot_iRAM_dist}
\title{Plot distribution of recovery time based on bootstrapped version of iRAM}
\usage{
plot_iRAM_dist(recovery.time.reps)
}
\arguments{
\item{recovery.time.reps}{bootstrapped version of recovery time}
}
\description{
Plot distribution of recovery time based on bootstrapped version of iRAM
}
\examples{
\dontshow{
data(bootstrap.iRAM)
plot_iRAM_dist(bootstrap.iRAM$recovery.time.reps)#'
}
\donttest{
data(bootstrap.iRAM)
plot_iRAM_dist(bootstrap.iRAM$recovery.time.reps)#'
}

}
