% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_time_profile}
\alias{plot_time_profile}
\title{Plot time profiles given a time-series generated by impulse response analysis}
\usage{
plot_time_profile(time.series.data, threshold = 0.01, xupper = 20)
}
\arguments{
\item{time.series.data}{data of impulse response in long format}

\item{threshold}{threshold of asymptote of equilibrium}

\item{xupper}{upper limit of x-axis}
}
\description{
Plot time profiles given a time-series generated by impulse response analysis
}
\examples{
\donttest{
# Similar to iRAM calculation, the corresponding plots of time profiles can be produced,
# once the time profiles data is calculated by impulse response analysis.

model.fit <- uSEM(var.number = 3,data = simts,lag.order = 1,verbose = FALSE,trim = TRUE)
beta.matrix <- parse.beta(var.number = 3, model.fit = model.fit, lag.order = 1, matrix = TRUE)
point.estimate.iRAM <- iRAM(model.fit,
                           beta = NULL,
                           var.number,
                           lag.order,
                           threshold = threshold,
                           boot = FALSE,
                           replication = replication,
                           steps= steps)

point.estimate.iRAM$recovery.time
plot_time_profile(time.series.data = point.estimate.iRAM$time.series.data, ,
                 threshold= .01,
                 xupper = 20)

# bootstrap version
boot.iRAM <- iRAM(model.fit,
                 beta = NULL,
                 var.number,
                 lag.order,
                 threshold = threshold,
                 boot = TRUE,
                 replication = replication,
                 steps= steps)


plot_time_profile(time.series.data = boot.iRAM$time.profile.data, ,
                 threshold= .01,
                 xupper = 20)
}

}
