% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_time_profile}
\alias{plot_time_profile}
\title{Plot time profiles given a time-series generated by impulse response analysis}
\usage{
plot_time_profile(time.series.data, var.number, threshold = 0.01,
  xupper = 20)
}
\arguments{
\item{time.series.data}{data of impulse response in long format}

\item{var.number}{number of variables in the time-series}

\item{threshold}{threshold of asymptote of equilibrium}

\item{xupper}{upper limit of x-axis}
}
\description{
Plot time profiles given a time-series generated by impulse response analysis
}
\examples{
\dontshow{
# Similar to iRAM calculation, the corresponding plots of time profiles can be produced,
# once the time profiles data is calculated by impulse response analysis.

data(bootstrap.iRAM)

plot_time_profile(time.series.data = bootstrap.iRAM$time.profile.data,
                 var.number = 3,
                 threshold= .01,
                 xupper = 20)
}

\donttest{
# Similar to iRAM calculation, the corresponding plots of time profiles can be produced,
# once the time profiles data is calculated by impulse response analysis.

data(model.fit)

point.estimate.iRAM <- iRAM(model.fit = model.fit,
                           beta = NULL,
                           var.number = 3,
                           lag.order = 1,
                           threshold = .01,
                           boot = FALSE,
                           replication = 200,
                           steps= 100)

point.estimate.iRAM$recovery.time
plot_time_profile(time.series.data = point.estimate.iRAM$time.series.data,
                 var.number = 3,
                 threshold= .01,
                 xupper = 20)

# bootstrap version
boot.iRAM <- iRAM(model.fit = model.fit,
                 beta = NULL,
                 var.number = 3,
                 lag.order = 1,
                 threshold = .01,
                 boot = TRUE,
    replication = 200,
    steps = 100
)


plot_time_profile(time.series.data = boot.iRAM$time.profile.data,
                 var.number = 3,
                 threshold= .01,
                 xupper = 20)
}

}
